app/requirements.txt
django==3.0.8
gunicorn==20.0.4
djangorestframework==3.11.0
django-cors-headers==3.4.0
djangorestframework-simplejwt==4.4.0
Pillow==7.2.0